backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.23.2
psycopg[binary]
